20.10.2023 | 12: 28

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: abostano <abostano@student.42istanbul.com  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 16:09:56 by abostano          #+#    #+#             */
-/*   Updated: 2023/10/13 16:13:45 by abostano         ###   ########.fr       */
+/*   Updated: 2023/10/19 10:44:33 by abostano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *a)
+size_t	ft_strlen(const char *a)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (a[i])
